Delete multiple task feature added
