+++ src/modules/UpdateIndex.js
@@ -1,0 +1,7 @@
+const UpdateIndex = (tasks) => {
+  tasks.forEach((task, index) => {
+    task.index = index + 1;
+  });
+};
+
+export default UpdateIndex;
