+++ src/modules/UpdateTasks.js
@@ -1,0 +1,3 @@
+const StoreTask = (tasks) => localStorage.setItem('tasks', JSON.stringify(tasks));
+
+export default StoreTask;
